+++ smart-contracts/wasm-transform/CHANGELOG.md
@@ -3,7 +3,11 @@
 ## Unreleased changes
 
 - `validate_module` is now parameterized by `ValidationConfig` which determines
-  which Wasm features should be allowed.
+  which Wasm features should be allowed. The currently supported configurable features are
+  - allow access to globals (defined in the current module) in data and element
+    initialization sections.
+  - allow instructions defined in the [sign extension operators](https://github.com/WebAssembly/sign-extension-ops/blob/master/proposals/sign-extension-ops/Overview.md)
+    Wasm proposal.
 
 ## concordium-wasm 2.0.0 (2023-06-16)